# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: xiaoan-platform API
  version: 2.0.0
  description: |
    xiaoan-platform backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/identity/admin/login:
    post:
      operationId: adminLogin
      tags: [identity]
      summary: Local admin login (Argon2id password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Admin session established
          content:
            application/json:
              schema:
                type: object
                required: [id, username, is_staff, auth_source]
                properties:
                  id: { type: integer }
                  username: { type: string }
                  is_staff: { type: boolean }
                  auth_source: { type: string }
                  display_name: { type: string }
                  display_id: { type: string }
        "401":
          description: invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ─────────────────────────────────────────────────────────────── auth ──
  /api/auth/login/:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Local username/password login (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in; also sets studio_session cookie
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
        "401":
          description: bad credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/auth/token/refresh/:
    post:
      operationId: tokenRefresh
      tags: [auth]
      summary: Rotate a JWT refresh token (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access: { type: string }
                  refresh: { type: string }
...
  /api/v2/runtimes/validate:
    post:
      operationId: validateRuntime
      tags: [runtimes]
      summary: Validate resource id format + remote visibility (best-effort)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider_key, runtime_type]
              properties:
                provider_key: { type: string }
                runtime_type: { type: string, default: agent }
                external_resource_id: { type: string }
                identity_mode: { type: string, default: user }
      responses:
        "200":
          description: Validation outcome (checked=false means skipped, not failed)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeValidationResult" }
        "400":
          description: field errors
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FieldErrors" }

  # ─────────────────────────────────────────────────────────────── runs ──
...
components:
...
  schemas:
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    SimpleError:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    FieldErrors:
      type: object
      description: map of field name -> list of messages
      additionalProperties:
        type: array
        items: { type: string }
    Tokens:
      type: object
      properties:
        access: { type: string }
        refresh: { type: string }
    User:
      type: object
      description: session_user_payload — the single canonical user shape
      required: [id, username]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string }
        avatar: { type: string }
        avatar_url: { type: string }
        auth_source: { type: string }
        is_staff: { type: boolean }
        display_name: { type: string }
        display_id: { type: string }
...
    RuntimeValidationResult:
      type: object
      required: [ok, checked, detail]
      properties:
        ok: { type: boolean }
        checked: { type: boolean }
        detail: { type: string }
...